refactor: move search test helpers into shared
