+++ shared/src/components/FileMatch.test.tsx
@@ -4,10 +4,10 @@ import * as React from 'react'
 import { cleanup, getAllByTestId, getByTestId, render } from '@testing-library/react'
 import _VisibilitySensor from 'react-visibility-sensor'
 import sinon from 'sinon'
-import { HIGHLIGHTED_FILE_LINES_REQUEST, NOOP_SETTINGS_CASCADE, RESULT } from '../../../web/src/search/testHelpers'
 import { MockVisibilitySensor } from './CodeExcerpt.test'
 import { FileMatch, IFileMatch } from './FileMatch'
 import { setLinkComponent } from './Link'
+import { HIGHLIGHTED_FILE_LINES_REQUEST, NOOP_SETTINGS_CASCADE, RESULT } from '../util/searchTestHelpers'
 
 jest.mock('react-visibility-sensor', (): typeof _VisibilitySensor => ({ children, onChange }) => (
     <>
